refactor: user controllers
