+++ package.json
@@ -7,7 +7,7 @@
     "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
-  "author": "",
+  "author": "@gogicha007",
   "license": "ISC",
   "description": "",
   "dependencies": {
@@ -16,8 +16,11 @@
     "uuid": "^10.0.0"
   },
   "devDependencies": {
+    "@types/jest": "^29.5.13",
     "@types/node": "^22.7.6",
+    "jest": "^29.7.0",
     "nodemon": "^3.1.7",
+    "ts-jest": "^29.2.5",
     "ts-node": "^10.9.2",
     "ts-node-dev": "^2.0.0",
     "typescript": "^5.6.3"
