+++ src/data/temp.js
@@ -1,0 +1,1 @@
+[{"username":"arut","age":21,"hobbies":["fishing","wining"],"id":"19dbf7e6-39ae-4f52-8fc0-8da1429b4917"},{"username":"jane","age":21,"hobbies":["fishing","wining"],"id":"2d812d95-e839-4a7e-8501-a46eb00134c0"},{"username":"jondo","age":52,"hobbies":["poems","swimming"],"id":"4682d479-4e16-49d4-921e-e1f43b537bf8"},{"id":"99d58f68-bf14-4a99-bc5d-2e17a18077a0","username":"merabi","age":27,"hobbies":["stone hurling","curling"]}]
