+++ src/routes/get.ts
@@ -1,14 +1,15 @@
 import { IncomingMessage, ServerResponse } from 'node:http';
-import { getAll, getById } from '../utils/getUsers';
+import { getAll, getById } from '../utils/userController';
+import { IFUser } from '../interfaces';
 
 
 export const getRequest = async (req: IncomingMessage, res: ServerResponse) => {
   switch (req.url) {
     case '/api/users':
-      const getUsers = (await getAll(req, res)) as string;
-      const result = getUsers ? getUsers : 'no data';
+      const getUsers = await getAll(req, res) as IFUser[];
+      const result = getUsers.length !==0 ? getUsers : 'no data';
       res.writeHead(200, { 'Content-Type': 'text/json' });
-      res.write(`Users: ${result}`);
+      res.write(`Users: ${JSON.stringify(result)}`);
       res.end();
       break;
     case (req.url as string).match(/\/api\/users\/\w+/)?.input:
